Numbers are now double

+++ src/main/java/org/mbruncic/cookingplanner/data/entity/RecipeIngredient.java
@@ -6,11 +6,11 @@ import javax.persistence.*;
 
 @Entity
 public class RecipeIngredient extends AbstractEntity {
-private int amount;
-public int getAmount() {
+private double amount;
+public double getAmount() {
     return amount;
 }
-public void setAmount(int amount) {
+public void setAmount(double amount) {
     this.amount = amount;
 }
